Added mask edge setup
Added the mask edge brighten setup. It multiplies the inverse of the mask of the grain and scattered grain by 1.8. Balancing the edge brightness.

Started adding the mask channels for the analyse and luminance but not fully working yet.

fixed a bug with external grain switch and the change depth nodes are now instances again